Move resume book to new controller
